--- a/code/builders/builder-vite/src/index.ts
+++ b/code/builders/builder-vite/src/index.ts
@@ -7,7 +7,6 @@ import type { Middleware, Options } from 'storybook/internal/types';
 import type { ViteDevServer } from 'vite';
 
 import { build as viteBuild } from './build';
-import { transformIframeHtml } from './transform-iframe-html';
 import type { ViteBuilder } from './types';
 import { createViteServer } from './vite-server';
 
@@ -17,25 +16,11 @@ export { hasVitePlugins } from './utils/has-vite-plugins';
 export * from './types';
 
 function iframeMiddleware(options: Options, server: ViteDevServer): Middleware {
-  return async (req, res, next) => {
-    if (!req.url || !req.url.match(/^\/iframe\.html($|\?)/)) {
-      next();
-      return;
-    }
-    // the base isn't used for anything, but it's required by the URL constructor
-    const url = new URL(req.url, 'http://localhost:6006');
-
-    // We need to handle `html-proxy` params for style tag HMR https://github.com/storybookjs/builder-vite/issues/266#issuecomment-1055677865
-    // e.g. /iframe.html?html-proxy&index=0.css
-    if (url.searchParams.has('html-proxy')) {
-      next();
-      return;
-    }
-
+  return async (req, res) => {
     const indexHtml = await readFile(require.resolve('@storybook/builder-vite/input/iframe.html'), {
       encoding: 'utf8',
     });
-    const transformed = await (await server).transformIndexHtml('/iframe.html', indexHtml);
+    const transformed = await server.transformIndexHtml('/iframe.html', indexHtml);
     res.setHeader('Content-Type', 'text/html');
     res.statusCode = 200;
     res.write(transformed);
@@ -46,7 +31,7 @@ function iframeMiddleware(options: Options, server: ViteDevServer): Middleware {
 let server: ViteDevServer;
 
 export async function bail(): Promise<void> {
-  return (await server)?.close();
+  return server?.close();
 }
 
 export const start: ViteBuilder['start'] = async ({