Closes https://github.com/storybookjs/storybook/issues/30378

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

- The preview build will now be awaited before Storybook's server starts so that all builder-specific routes and middlewares are registered
- I've added additional packages to Vite's `optimizeDeps.include` configuration to avoid issues when Storybook starts up the first time.

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-30567-sha-f5452a43`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-30567-sha-f5452a43 sandbox` or in an existing project with `npx storybook@0.0.0-pr-30567-sha-f5452a43 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-30567-sha-f5452a43`](https://npmjs.com/package/storybook/v/0.0.0-pr-30567-sha-f5452a43) |
| **Triggered by** | @valentinpalkovic |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`valentin/fix-runtime-iframe-error-on-first-load`](https://github.com/storybookjs/storybook/tree/valentin/fix-runtime-iframe-error-on-first-load) |
| **Commit** | [`f5452a43`](https://github.com/storybookjs/storybook/commit/f5452a43c2e23360552688146112dc7a2a25fb63) |
| **Datetime** | Wed Feb 19 10:52:58 UTC 2025 (`1739962378`) |
| **Workflow run** | [13410852699](https://github.com/storybookjs/storybook/actions/runs/13410852699) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=30567`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  80.5 MB | 80.5 MB | 29 kB | **-1.4** | 0% |
| initSize |  80.5 MB | 80.5 MB | 29 kB | **-1.4** | 0% |
| diffSize |  97 B | 97 B | 0 B | - | 0% |
| buildSize |  7.31 MB | 7.31 MB | -395 B | **-1.58** | 0% |
| buildSbAddonsSize |  1.9 MB | 1.9 MB | 380 B | 0.23 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.88 MB | 1.88 MB | -692 B | -0.23 | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.97 MB | 3.97 MB | -312 B | -0.23 | 0% |
| buildPreviewSize |  3.34 MB | 3.34 MB | -83 B | **-18.76** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  15.9s | 26.9s | 11s | **1.79** | 🔺40.9% |
| generateTime |  20.8s | 18.6s | -2s -250ms | -0.86 | -12.1% |
| initTime |  4.7s | 4.2s | -532ms | -1.15 | -12.7% |
| buildTime |  10.5s | 8.5s | -2s -37ms | -0.99 | -23.9% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.1s | 4.6s | -519ms | **-1.5** | 🔰-11.3% |
| devManagerResponsive |  3.8s | 4.4s | 562ms | 0.57 | 12.7% |
| devManagerHeaderVisible |  871ms | 757ms | -114ms | -0.37 | -15.1% |
| devManagerIndexVisible |  907ms | 768ms | -139ms | -0.49 | -18.1% |
| devStoryVisibleUncached |  4.1s | 1.8s | -2s -369ms | **-4.29** | 🔰-131.3% |
| devStoryVisible |  907ms | 786ms | -121ms | -0.57 | -15.4% |
| devAutodocsVisible |  841ms | 755ms | -86ms | -0.34 | -11.4% |
| devMDXVisible |  823ms | 683ms | -140ms | -1.1 | -20.5% |
| buildManagerHeaderVisible |  683ms | 602ms | -81ms | **-1.65** | 🔰-13.5% |
| buildManagerIndexVisible |  695ms | 630ms | -65ms | **-1.66** | 🔰-10.3% |
| buildStoryVisible |  667ms | 590ms | -77ms | **-1.54** | 🔰-13.1% |
| buildAutodocsVisible |  620ms | 543ms | -77ms | -0.78 | -14.2% |
| buildMDXVisible |  571ms | 519ms | -52ms | **-1.34** | 🔰-10% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Added new dependencies to Vite's optimizeDeps configuration to prevent runtime and iframe 404 errors that occur on first load of Storybook.

- Added `@storybook/addon-interactions/preview` to `INCLUDE_CANDIDATES` in `code/builders/builder-vite/src/optimizeDeps.ts`
- Added `react/jsx-dev-runtime` to `INCLUDE_CANDIDATES` for proper React development runtime support
- Ensures Vite pre-bundles CommonJS dependencies that need ESM conversion
- Fixes startup issues reported in versions after Storybook 8.3.5



<!-- /greptile_comment -->